fix client side error
